apps/agent/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
openai>=1.0.0
numpy>=1.24.3
scipy>=1.11.0
scikit-learn>=1.3.0